+++ Storefront/Includes/deletestore_inc.php
@@ -21,25 +21,27 @@ if(isset($_POST['delete_store'])) { //User got here legitimately
         if($row = mysqli_fetch_assoc($result)) { //Puts results into array
             $img_dest = '../Uploads/Store/'.$row['img_dest'];
 
-            if(!unlink($img_dest)) { //If image doesn't delete
-                header('Location: ../manage.php?error=notdeleted');
+            if($img_dest != '../Uploads/Store/default.jpg')
+            {
+                if(!unlink($img_dest)) { //If image doesn't delete
+                    header('Location: ../manage.php?error=notdeleted');
+                    exit();
+                }
+            }
+
+            $sql = 'DELETE FROM stores WHERE store_id=?;';
+            $statement = mysqli_stmt_init($conn);
+
+            if(!mysqli_stmt_prepare($statement, $sql)) { //If connection fails
+                header('Location: ../manage.php?error=sqlerror');
                 exit();
             }
             else {
-                $sql = 'DELETE FROM stores WHERE store_id=?;';
-                $statement = mysqli_stmt_init($conn);
+                mysqli_stmt_bind_param($statement, 'i', $store_id);
+                mysqli_stmt_execute($statement);
 
-                    if(!mysqli_stmt_prepare($statement, $sql)) { //If connection fails
-                    header('Location: ../manage.php?error=sqlerror');
-                    exit();
-                    }
-                    else {
-                        mysqli_stmt_bind_param($statement, 'i', $store_id);
-                        mysqli_stmt_execute($statement);
-
-                        header('Location: ../manage.php?delete=success');
-                        exit();
-                    }
+                header('Location: ../manage.php?delete=success');
+                exit();
             }
         }
         else {
